# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Synorlab Interviewer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface UploadJDBody {
  /** Raw job description text */
  text: string;
}

export interface ParsedJD {
  id: number;
  role: string;
  company: string;
  skills: string[];
  experienceLevel: string;
  responsibilities: string[];
  rawText: string;
  createdAt: string;
}

export interface CreateInterviewBody {
  /** ID of the parsed job description */
  jdId: number;
}

export type InterviewStatus =
  (typeof InterviewStatus)[keyof typeof InterviewStatus];

export const InterviewStatus = {
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface Interview {
  id: number;
  userId: string;
  jdId: number;
  status: InterviewStatus;
  createdAt: string;
}

export type InterviewSummaryStatus =
  (typeof InterviewSummaryStatus)[keyof typeof InterviewSummaryStatus];

export const InterviewSummaryStatus = {
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface InterviewSummary {
  id: number;
  jdId: number;
  role: string;
  status: InterviewSummaryStatus;
  score?: number | null;
  messageCount: number;
  createdAt: string;
}

export type InterviewMessageRole =
  (typeof InterviewMessageRole)[keyof typeof InterviewMessageRole];

export const InterviewMessageRole = {
  ai: "ai",
  user: "user",
} as const;

export interface InterviewMessage {
  id: number;
  role: InterviewMessageRole;
  content: string;
  createdAt: string;
}

export type InterviewDetailStatus =
  (typeof InterviewDetailStatus)[keyof typeof InterviewDetailStatus];

export const InterviewDetailStatus = {
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface CriteriaScore {
  dimension: string;
  score: number;
  weight: number;
  feedback: string;
}

export interface QuestionEvaluation {
  question: string;
  answer: string;
  score: number;
  feedback: string;
  idealAnswer: string;
}

export interface Evaluation {
  id: number;
  interviewId: number;
  overallScore: number;
  feedback: string;
  strengths: string[];
  improvements: string[];
  criteriaScores: CriteriaScore[];
  questionEvals: QuestionEvaluation[];
  createdAt: string;
}

export interface InterviewDetail {
  id: number;
  jdId: number;
  role: string;
  company: string;
  status: InterviewDetailStatus;
  messages: InterviewMessage[];
  evaluation?: Evaluation | null;
  createdAt: string;
}

export interface RespondToInterviewBody {
  content: string;
}

export interface AIMessage {
  id: number;
  content: string;
  isComplete: boolean;
}

export interface InterviewStats {
  totalInterviews: number;
  completedInterviews: number;
  averageScore: number;
  recentInterviews: InterviewSummary[];
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  student: "student",
  admin: "admin",
  facility: "facility",
} as const;

export interface UserProfile {
  id: string;
  email: string;
  role: UserProfileRole;
  createdAt: string;
}

export type UpdateUserProfileBodyRole =
  (typeof UpdateUserProfileBodyRole)[keyof typeof UpdateUserProfileBodyRole];

export const UpdateUserProfileBodyRole = {
  student: "student",
  admin: "admin",
  facility: "facility",
} as const;

export interface UpdateUserProfileBody {
  role: UpdateUserProfileBodyRole;
}

export interface AdminUser {
  id: string;
  email: string;
  role: string;
  totalInterviews: number;
  createdAt: string;
}

export interface AdminStats {
  totalUsers: number;
  totalInterviews: number;
  completedInterviews: number;
  averageScore: number;
}

export type AdminUpdateUserRoleBodyRole =
  (typeof AdminUpdateUserRoleBodyRole)[keyof typeof AdminUpdateUserRoleBodyRole];

export const AdminUpdateUserRoleBodyRole = {
  student: "student",
  admin: "admin",
  facility: "facility",
} as const;

export interface AdminUpdateUserRoleBody {
  role: AdminUpdateUserRoleBodyRole;
}

export type AdminInterviewSummaryStatus =
  (typeof AdminInterviewSummaryStatus)[keyof typeof AdminInterviewSummaryStatus];

export const AdminInterviewSummaryStatus = {
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface AdminInterviewSummary {
  id: number;
  userId: string;
  userEmail: string;
  role: string;
  status: AdminInterviewSummaryStatus;
  score?: number | null;
  messageCount: number;
  createdAt: string;
}

export type InviteRole = (typeof InviteRole)[keyof typeof InviteRole];

export const InviteRole = {
  student: "student",
  admin: "admin",
  facility: "facility",
} as const;

export interface Invite {
  id: number;
  email: string;
  role: InviteRole;
  token: string;
  used: boolean;
  createdAt: string;
}

export type CreateInviteBodyRole =
  (typeof CreateInviteBodyRole)[keyof typeof CreateInviteBodyRole];

export const CreateInviteBodyRole = {
  student: "student",
  admin: "admin",
  facility: "facility",
} as const;

export interface CreateInviteBody {
  email: string;
  role: CreateInviteBodyRole;
}

export type AccessCodeRole =
  (typeof AccessCodeRole)[keyof typeof AccessCodeRole];

export const AccessCodeRole = {
  student: "student",
  admin: "admin",
  facility: "facility",
} as const;

export interface AccessCode {
  id: number;
  code: string;
  role: AccessCodeRole;
  active: boolean;
  maxUses: number;
  usedCount: number;
  createdAt: string;
}

export type CreateAccessCodeBodyRole =
  (typeof CreateAccessCodeBodyRole)[keyof typeof CreateAccessCodeBodyRole];

export const CreateAccessCodeBodyRole = {
  student: "student",
  admin: "admin",
  facility: "facility",
} as const;

export interface CreateAccessCodeBody {
  role: CreateAccessCodeBodyRole;
  code?: string;
  maxUses?: number;
}

export interface RedeemCodeBody {
  code: string;
}

export interface FacilityStudent {
  id: string;
  email: string;
  totalInterviews: number;
  completedInterviews: number;
  averageScore?: number | null;
  lastActive?: string | null;
  createdAt: string;
}

export interface FacilityStats {
  totalStudents: number;
  activeThisWeek: number;
  averageScore: number;
  completionRate: number;
}