# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Synorlab Interviewer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get the current user's profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateUserProfile
      tags: [users]
      summary: Update user profile (role)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileBody"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/redeem-code:
    post:
      operationId: redeemAccessCode
      tags: [users]
      summary: Redeem an access code to change role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeemCodeBody"
      responses:
        "200":
          description: Role updated via access code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /admin/users/{userId}/role:
    put:
      operationId: updateUserRole
      tags: [admin]
      summary: Update a user's role (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserRoleBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/users/{userId}:
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete a user and all their data (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /admin/interviews/{id}:
    delete:
      operationId: deleteAdminInterview
      tags: [admin]
      summary: Delete an interview (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Interview deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/invites:
    get:
      operationId: listInvites
      tags: [admin]
      summary: List all email invites (admin only)
      responses:
        "200":
          description: List of invites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invite"
    post:
      operationId: createInvite
      tags: [admin]
      summary: Create an email invite (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteBody"
      responses:
        "201":
          description: Invite created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/invites/{id}:
    delete:
      operationId: deleteInvite
      tags: [admin]
      summary: Delete an invite (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invite deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/access-codes:
    get:
      operationId: listAccessCodes
      tags: [admin]
      summary: List all access codes (admin only)
      responses:
        "200":
          description: List of access codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccessCode"
    post:
      operationId: createAccessCode
      tags: [admin]
      summary: Create an access code (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccessCodeBody"
      responses:
        "201":
          description: Access code created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessCode"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/access-codes/{id}:
    delete:
      operationId: deleteAccessCode
      tags: [admin]
      summary: Delete an access code (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Access code deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin, facility]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
        - createdAt
    UpdateUserProfileBody:
      type: object
      properties:
        role:
          type: string
          enum: [student, admin, facility]
      required:
        - role
    AdminUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
        totalInterviews:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
        - totalInterviews
        - createdAt
...
    AdminUpdateUserRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [student, admin, facility]
      required:
        - role
...
    Invite:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
          enum: [student, admin, facility]
        token:
          type: string
        used:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
        - token
        - used
        - createdAt
    CreateInviteBody:
      type: object
      properties:
        email:
          type: string
        role:
          type: string
          enum: [student, admin, facility]
      required:
        - email
        - role
    AccessCode:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        role:
          type: string
          enum: [student, admin, facility]
        active:
          type: boolean
        maxUses:
          type: integer
        usedCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - code
        - role
        - active
        - maxUses
        - usedCount
        - createdAt
    CreateAccessCodeBody:
      type: object
      properties:
        role:
          type: string
          enum: [student, admin, facility]
        code:
          type: string
        maxUses:
          type: integer
      required:
        - role
    RedeemCodeBody:
      type: object
      properties:
        code:
          type: string
      required:
        - code
...