# lib/api-zod/src/generated/types/accessCode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Synorlab Interviewer API
 * OpenAPI spec version: 0.1.0
 */
import type { AccessCodeRole } from "./accessCodeRole";

export interface AccessCode {
  id: number;
  code: string;
  role: AccessCodeRole;
  active: boolean;
  maxUses: number;
  usedCount: number;
  createdAt: Date;
}