# lib/api-zod/src/generated/types/createAccessCodeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Synorlab Interviewer API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAccessCodeBodyRole } from "./createAccessCodeBodyRole";

export interface CreateAccessCodeBody {
  role: CreateAccessCodeBodyRole;
  code?: string;
  maxUses?: number;
}