# lib/api-zod/src/generated/types/createInviteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Synorlab Interviewer API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateInviteBodyRole } from "./createInviteBodyRole";

export interface CreateInviteBody {
  email: string;
  role: CreateInviteBodyRole;
}