# lib/api-zod/src/generated/types/invite.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Synorlab Interviewer API
 * OpenAPI spec version: 0.1.0
 */
import type { InviteRole } from "./inviteRole";

export interface Invite {
  id: number;
  email: string;
  role: InviteRole;
  token: string;
  used: boolean;
  createdAt: Date;
}